Move batchAdd into server/search/utils.js
